fix Markdown links
