+++ apps/web/app/routes/Home/route.tsx
@@ -1,124 +1,32 @@
 import type { MetaFunction } from "react-router"
-import { Link, useFetcher, useRouteLoaderData } from "react-router"
+import { Link } from "react-router"
 
 import { ErrorBoundary } from "@sentry/react"
-import marked from "marked"
-import type { ComponentProps, JSX, ReactNode } from "react"
-import { Suspense, useMemo } from "react"
-import ReactRelay, { useLazyLoadQuery } from "react-relay"
+import type { ReactNode } from "react"
+import ReactRelay from "react-relay"
 import { Card } from "~/components/Card"
 import { LayoutBody, LayoutPane } from "~/components/Layout"
 import {
 	List,
 	ListItem,
-	ListItemAvatar,
 	ListItemContent,
-	ListItemImg,
 	ListItemContentSubtitle as ListItemSubtitle,
 	ListItemContentTitle as ListItemTitle,
 } from "~/components/List"
 
 import { client_operation } from "~/lib/client"
 
-import { route_media, route_user } from "~/lib/route"
-
 // console.log(R)
 
-import { MediaCover } from "~/lib/entry/MediaCover"
-
-import createDOMPurify from "dompurify"
-import {
-	TooltipRich,
-	TooltipRichActions,
-	TooltipRichContainer,
-	TooltipRichTrigger,
-} from "~/components/Tooltip"
-
-import { Button } from "~/components/Button"
-import { type clientLoader as rootLoader } from "~/root"
-
-import type { routeMediaCardQuery } from "~/gql/routeMediaCardQuery.graphql"
 import type { routeNavFeedMediaQuery } from "~/gql/routeNavFeedMediaQuery.graphql"
 import type { routeNavFeedQuery } from "~/gql/routeNavFeedQuery.graphql"
-import type { routeUserCardQuery } from "~/gql/routeUserCardQuery.graphql"
 import { loadQuery, usePreloadedQuery } from "~/lib/Network"
-import { numberToString } from "~/lib/numberToString"
-import { m } from "~/lib/paraglide"
 import * as Predicate from "~/lib/Predicate"
 import { getThemeFromHex } from "~/lib/theme"
-import type { clientAction as userFollowAction } from "../UserFollow/route"
 import type { Route } from "./+types/route"
+import { Markdown } from "./Markdown"
+import { options } from "./options"
 const { graphql } = ReactRelay
-
-function MediaLink({
-	mediaId,
-	type,
-	slug,
-	...props
-}: Omit<ComponentProps<typeof Link>, "to"> & {
-	mediaId: number
-	type: string
-	slug?: string
-}) {
-	return (
-		<Link to={route_media({ id: mediaId })} {...props}>
-			<ErrorBoundary
-				fallback={<>{slug ? `${type}${slug}` : `Failed to load ${type}`}</>}
-			>
-				<Suspense fallback="Loading...">
-					<MediaCard mediaId={mediaId} type={type}></MediaCard>
-				</Suspense>
-			</ErrorBoundary>
-		</Link>
-	)
-}
-
-function MediaCard(props: { mediaId: number; type: string }) {
-	const media = useLazyLoadQuery<routeMediaCardQuery>(
-		graphql`
-			query routeMediaCardQuery($id: Int) {
-				Media(id: $id) {
-					title {
-						userPreferred
-					}
-					coverImage {
-						theme
-					}
-					...MediaCover_media
-				}
-			}
-		`,
-		{
-			id: props.mediaId,
-		}
-	).Media
-
-	return (
-		media && (
-			<Card
-				className={`not-prose contrast-standard theme-light contrast-more:contrast-high dark:theme-dark inline-flex overflow-hidden p-0 text-start`}
-				style={media.coverImage?.theme ?? undefined}
-				render={<span />}
-			>
-				<List className="p-0" render={<span />}>
-					<ListItem render={<span />}>
-						<ListItemImg render={<span></span>}>
-							<MediaCover media={media} />
-						</ListItemImg>
-						<ListItemContent render={<span></span>}>
-							<ListItemTitle render={<span />}>
-								{media.title?.userPreferred}
-							</ListItemTitle>
-							<ListItemSubtitle render={<span />}>
-								{props.type}
-							</ListItemSubtitle>
-						</ListItemContent>
-					</ListItem>
-				</List>
-			</Card>
-		)
-	)
-}
 
 function matchMediaId(s: string) {
 	return [...s.matchAll(/https:\/\/anilist.co\/(anime|manga)\/(\d+)/g)]
@@ -272,7 +180,9 @@ export default function Index({ loaderData }: Route.ComponentProps): ReactNode {
 												</ListItem>
 											</List>
 											<ErrorBoundary fallback={<>Failed to parse markdown</>}>
-												{activity.text && <Markdown>{activity.text}</Markdown>}
+												{activity.text && (
+													<Markdown options={options}>{activity.text}</Markdown>
+												)}
 											</ErrorBoundary>
 										</Card>
 									</li>
@@ -288,433 +198,6 @@ export default function Index({ loaderData }: Route.ComponentProps): ReactNode {
 				</div> */}
 			</LayoutPane>
 		</LayoutBody>
-	)
-}
-
-export interface Options {
-	replace: Partial<{
-		[K in keyof JSX.IntrinsicElements]: (
-			props: JSX.IntrinsicElements[K]
-		) => ReactNode
-	}>
-}
-
-const options = {
-	replace: {
-		center(props) {
-			return <center {...props} />
-		},
-		p(props) {
-			return <p {...props} />
-		},
-		a(props) {
-			if (!props.href?.trim()) {
-				return <a className="text-primary">{props.children}</a>
-			}
-
-			let mediaId: number
-
-			if (
-				props.className === "media-link" &&
-				"data-id" in props &&
-				"data-type" in props &&
-				"data-slug" in props &&
-				Predicate.isString(props["data-type"]) &&
-				Predicate.isString(props["data-slug"]) &&
-				isFinite((mediaId = Number(props["data-id"])))
-			) {
-				return (
-					<MediaLink
-						{...props}
-						mediaId={mediaId}
-						type={props["data-type"]}
-						slug={props["data-slug"]}
-					/>
-				)
-			}
-
-			if (
-				"data-user-name" in props &&
-				Predicate.isString(props["data-user-name"])
-			) {
-				return (
-					<UserLink userName={props["data-user-name"]}>
-						{props.children}
-					</UserLink>
-				)
-			}
-
-			return (
-				<a
-					{...props}
-					rel="noopener noreferrer"
-					target="_blank"
-					className="text-primary"
-				>
-					{props.children}
-				</a>
-			)
-		},
-	},
-} satisfies Options
-
-function UserCard(props: { userName: string }) {
-	const data = useLazyLoadQuery<routeUserCardQuery>(
-		graphql`
-			query routeUserCardQuery($userName: String!) {
-				User(name: $userName) {
-					id
-					avatar {
-						medium
-						large
-					}
-					isFollower
-					isFollowing
-				}
-			}
-		`,
-		{ userName: props.userName }
-	)
-
-	const rootData = useRouteLoaderData<typeof rootLoader>("root")
-	const follow = useFetcher<typeof userFollowAction>({
-		key: `${props.userName}-follow`,
-	})
-
-	return (
-		data.User && (
-			<>
-				<div className="-mx-4 -my-2">
-					<List lines={"two"} className="">
-						<ListItem className="hover:state-none">
-							<ListItemAvatar>
-								{data.User.avatar?.large && (
-									<img
-										src={data.User.avatar.large}
-										className="bg-(image:--bg) bg-cover bg-center object-cover object-center"
-										style={{
-											"--bg": `url(${data.User.avatar.medium ?? ""})`,
-										}}
-										loading="lazy"
-										alt=""
-									/>
-								)}
-							</ListItemAvatar>
-							<ListItemContent>
-								<ListItemTitle>{props.userName}</ListItemTitle>
-								<ListItemSubtitle>
-									{data.User.isFollower ? "Follower" : "Not follower"}
-								</ListItemSubtitle>
-							</ListItemContent>
-						</ListItem>
-					</List>
-				</div>
-
-				<TooltipRichActions>
-					{rootData?.Viewer?.name &&
-						rootData.Viewer.name !== props.userName && (
-							<follow.Form
-								method="post"
-								action={`/user/${numberToString(data.User.id)}/follow`}
-							>
-								<input
-									type="hidden"
-									name="isFollowing"
-									value={
-										(follow.formData?.get("isFollowing") ??
-										follow.data?.ToggleFollow.isFollowing ??
-										data.User.isFollowing)
-											? ""
-											: "true"
-									}
-									id=""
-								/>
-
-								<Button type="submit" aria-disabled={!data.User.id}>
-									{(follow.formData?.get("isFollowing") ??
-									follow.data?.ToggleFollow.isFollowing ??
-									data.User.isFollowing)
-										? m.unfollow_button()
-										: m.follow_button()}
-								</Button>
-							</follow.Form>
-						)}
-				</TooltipRichActions>
-				{/* <TooltipRichSubhead>{props.children}</TooltipRichSubhead>
-			<TooltipRichSupportingText>{props.children}</TooltipRichSupportingText> */}
-			</>
-		)
-	)
-}
-
-function UserLink(props: { userName: string; children: ReactNode }) {
-	return (
-		<TooltipRich placement="top">
-			<TooltipRichTrigger
-				render={
-					<Link to={route_user({ userName: props.userName })}>
-						{props.children}
-					</Link>
-				}
-			/>
-
-			<ErrorBoundary>
-				<Suspense>
-					<TooltipRichContainer className="text-start" unmountOnHide>
-						<UserCard userName={props.userName}></UserCard>
-					</TooltipRichContainer>
-				</Suspense>
-			</ErrorBoundary>
-		</TooltipRich>
-	)
-}
-
-// const options: HTMLReactParserOptions = {
-// 	replace(domNode) {
-// 		if (
-// 			domNode instanceof Element &&
-// 			domNode.name === "a" &&
-// 			!domNode.attribs["href"]?.trim()
-// 		) {
-// 			return (
-// 				<span className="text-primary">
-// 					{domToReact(domNode.children, options)}
-// 				</span>
-// 			)
-// 		}
-
-// 		if (domNode instanceof Element && domNode.name === "center") {
-// 			return (
-// 				<span className="text-center">
-// 					{domToReact(domNode.children, options)}
-// 				</span>
-// 			)
-// 		}
-
-// 		if (
-// 			domNode instanceof Element &&
-// 			domNode.attribs["class"] === "media-link" &&
-// 			domNode.attribs["data-id"] &&
-// 			domNode.name === "a"
-// 		) {
-// 			return <MediaLink mediaId={domNode.attribs["data-id"]}></MediaLink>
-// 		}
-
-// 		if (
-// 			domNode instanceof Element &&
-// 			domNode.attribs["class"] === "user-link" &&
-// 			domNode.attribs["data-user-name"] &&
-// 			domNode.name === "a"
-// 		) {
-// 			return (
-// 				<Link  to={route_user({ userName: domNode.attribs["data-user-name"] })}>
-// 					{domToReact(domNode.children, options)}
-// 				</Link>
-// 			)
-// 		}
-
-// 		if (domNode instanceof Element && domNode.name === "a") {
-// 			return (
-// 				<a
-// 					{...attributesToProps(domNode.attribs)}
-// 					rel="noopener noreferrer"
-// 					target="_blank"
-// 				>
-// 					{domToReact(domNode.children, options)}
-// 				</a>
-// 			)
-// 		}
-// 	}
-// }
-
-function Markdown(props: { children: string }) {
-	return (
-		<div className="prose md:prose-lg lg:prose-xl dark:prose-invert prose-img:rounded-md prose-video:rounded-md max-w-full overflow-x-auto">
-			{useMemo(
-				() => parse(markdownHtml(props.children), options),
-				[props.children]
-			)}
-		</div>
-	)
-}
-
-function getAttributes(attributes: Record<string, string>) {
-	const {
-		class: className,
-		allowfullscreen: allowFullScreen,
-		frameborder: frameBorder,
-		..._attributes
-	} = attributes
-
-	return {
-		..._attributes,
-		...(className ? { className } : {}),
-		...(allowFullScreen ? { allowFullScreen } : {}),
-		...(frameBorder ? { frameBorder } : {}),
-	}
-}
-
-function traverse(element: ChildNode, options: Options): ReactNode {
-	if (element.nodeType === Node.TEXT_NODE) {
-		return element.textContent
-	} else if (element instanceof HTMLElement) {
-		const Name =
-			element.tagName.toLowerCase() as unknown as keyof JSX.IntrinsicElements
-		const attributes: Record<string, string> = {}
-		for (const attribute of element.attributes) {
-			attributes[attribute.name] = attribute.value
-		}
-
-		const fullProps: object = {
-			...getAttributes(attributes),
-			children: traverseCollection(element.childNodes, options),
-		}
-
-		if (Name in options.replace) {
-			return options.replace[Name]?.(fullProps)
-		}
-
-		return <Name {...fullProps} />
-	}
-}
-
-function traverseCollection(
-	children: NodeListOf<ChildNode>,
-	options: Options
-): ReactNode {
-	return Array.from(children).reduce<ReactNode>(
-		(acc, node) => (
-			<>
-				{acc}
-				{traverse(node, options)}
-			</>
-		),
-		null
-	)
-}
-
-function parse(html: string, options: Options): ReactNode {
-	const subdocument = new DOMParser().parseFromString(html, "text/html")
-
-	return traverseCollection(subdocument.body.childNodes, options)
-}
-
-function sanitizeHtml(t: string) {
-	const DOMPurify = createDOMPurify(window)
-
-	DOMPurify.addHook("afterSanitizeAttributes", (node) => {
-		if ("target" in node) {
-			node.setAttribute("target", "_blank")
-			node.setAttribute("rel", "noopener noreferrer")
-		}
-	})
-
-	const out: string = DOMPurify.sanitize(t, {
-		ALLOWED_TAGS: [
-			"a",
-			"b",
-			"blockquote",
-			"br",
-			"center",
-			"del",
-			"div",
-			"em",
-			"font",
-			"h1",
-			"h2",
-			"h3",
-			"h4",
-			"h5",
-			"hr",
-			"i",
-			"img",
-			"li",
-			"ol",
-			"p",
-			"pre",
-			"code",
-			"span",
-			"strike",
-			"strong",
-			"ul",
-		],
-		ALLOWED_ATTR: ["align", "height", "href", "src", "target", "width", "rel"],
-	})
-
-	return out
-}
-
-function markdownHtml(t: string) {
-	marked.setOptions({
-		gfm: true,
-		tables: true,
-		breaks: true,
-		pedantic: false,
-		sanitize: false,
-		smartLists: true,
-		smartypants: false,
-	})
-
-	const d = new marked.Renderer(),
-		u = new marked.Lexer()
-
-	d.link = (t: string, e: string, a: string) => {
-		return `<a href="${t}" title="${e}" target="_blank" rel="noopener noreferrer">${a}</a>`
-	}
-	u.rules.heading = /^ *(#{1,6}) *([^\n]+?) *#* *(?:\n+|$)/
-
-	return (
-		(t = t.replace(
-			/(http)(:([/|.|\w|\s|-])*\.(?:jpg|.jpeg|gif|png|mp4|webm))/gi,
-			"$1s$2"
-		)),
-		(t = t.replaceAll(
-			/(^|>| )@([A-Za-z0-9]+)/gm,
-			(group, one: string, userName: string) => {
-				return `${one}<a class='user-link' data-user-name='${userName}' href='${route_user({ userName: userName })}'>@${userName}</a>`
-			}
-		)),
-		(t = t.replace(
-			/img\s?(\d+%?)?\s?\((.[\S]+)\)/gi,
-			"<img width='$1' src='$2'>"
-		)),
-		(t = t.replace(
-			/youtube\s?\([^]*?([-_0-9A-Za-z]{10,15})[^]*?\)/gi,
-			"youtube ($1)"
-		)),
-		(t = t.replace(
-			/webm\s?\(h?([A-Za-z0-9-._~:/?#[\]@!$&()*+,;=%]+)\)/gi,
-			"webmv(`$1`)"
-		)),
-		(t = t.replace(/~{3}([^]*?)~{3}/gm, "+++$1+++")),
-		(t = t.replace(/~!([^]*?)!~/gm, '<div rel="spoiler">$1</div>')),
-		(t = sanitizeHtml(
-			// @ts-expect-error marked error types are not correct
-			marked(t, {
-				renderer: d,
-				lexer: u,
-			})
-		)),
-		(t = t.replace(/\+{3}([^]*?)\+{3}/gm, "<center>$1</center>")),
-		(t = t.replace(
-			/<div rel="spoiler">([\s\S]*?)<\/div>/gm,
-			"<details><summary>Show spoiler</summary>$1</details>"
-		)),
-		(t = t.replace(
-			/youtube\s?\(([-_0-9A-Za-z]{10,15})\)/gi,
-			"<iframe width='450' height='315' src='https://www.youtube.com/embed/$1/' title='YouTube video player' frameBorder='0' name='1' class='rounded-md' allow='accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share;' allowfullscreen loading='lazy'>Your browser does not support iframes</iframe>"
-		)),
-		(t = t.replace(
-			/webmv\s?\(<code>([A-Za-z0-9-._~:/?#[\]@!$&()*+,;=%]+)<\/code>\)/gi,
-			"<video muted loop controls><source src='h$1' type='video/webm'>Your browser does not support the video tag.</video>"
-		)),
-		(t = t.replace(
-			/(?:<a.*href="https?:\/\/anilist.co\/(anime|manga)\/)([0-9]+)(\/\w+)?.*?>(?:https?:\/\/anilist.co\/(?:anime|manga)\/[0-9]+).*?<\/a>/gm,
-			(group, type: string, mediaId: string, slug: string) => {
-				return `<a class="media-link" href="${route_media({ id: Number(mediaId) })}" data-type="${type}" data-slug="${slug}" data-id="${mediaId}">Loading...</a>`
-			}
-		)),
-		t
 	)
 }
